pyproject.toml
[project]
name = "aws-mobility-elt-pipeline"
version = "0.1.0"
dependencies = [
  "apache-airflow>=2.8.0",
  "apache-airflow-providers-amazon>=9.19.0",
  "apache-airflow-task-sdk>=1.1.6",
  "boto3>=1.42.22",
  "great-expectations>=1.10.0",
  "pyiceberg>=0.10.0",
  "pyspark==3.5.3",
  "pytest>=9.0.2",
  "pytest-spark>=0.8.0",
  "pyyaml>=6.0.3",
  "requests>=2.32.5",
  "ruff>=0.14.11",
  "structlog>=25.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
